update README.md File

+++ src/test/java/com/appium/test/util/GeneralUtil.java
@@ -1,5 +1,5 @@
 package com.appium.test.util;
 
 public class GeneralUtil {
-    public static final int WAIT_TIME = 30;
+    public static final long WAIT_TIME = 30;
 }
